+++ libs/search-widget/src/tiles/base-tile/DocumentTile.svelte
@@ -16,6 +16,7 @@
   import { hasViewerSearchError, viewerSearchQuery, viewerSearchResults } from '../../core/stores/viewer-search.store';
   import { navigateToLink } from '../../core/stores/widget.store';
   import TileHeader from './TileHeader.svelte';
+  import { resource } from '../../core/stores/resource.store';
 
   export let result: Search.SmartResult = { id: '' } as Search.SmartResult;
   export let resourceObs: Observable<Resource>;
@@ -44,7 +45,6 @@
   $: isMobile = innerWidth < 448;
   $: defaultTransitionDuration = expanded ? Duration.MODERATE : 0;
 
-  let resource: Resource | undefined;
   let paragraphList: WidgetParagraph[];
   const isSearchingInResource = new BehaviorSubject(false);
   const matchingParagraphs$: Observable<WidgetParagraph[]> = combineLatest([
@@ -87,8 +87,9 @@
       viewerSearchQuery.set(globalQuery.getValue());
       expanded = true;
       freezeBackground(true);
-      if (!resource) {
-        resourceObs.subscribe((res) => (resource = res));
+
+      if (!resource.value) {
+        resourceObs.subscribe((res) => resource.set(res));
       }
     }
     setTimeout(() => setHeaderActionWidth());
@@ -142,9 +143,10 @@
 
   const findInPdf = () => {
     const query = viewerSearchQuery.getValue();
-    if (resource && query) {
+    const currentResource = resource.value;
+    if (currentResource && query) {
       isSearchingInResource.next(true);
-      resource
+      currentResource
         .search(query, [Search.ResourceFeatures.PARAGRAPH], { highlight: true })
         .pipe(
           tap((results) => {