+++ libs/search-widget/src/tiles/base-tile/TileHeader.svelte
@@ -2,8 +2,10 @@
   import DocTypeIndicator from '../../common/indicators/DocTypeIndicator.svelte';
   import { IconButton } from '../../common';
   import { _ } from '../../core/i18n';
-  import { Search } from '@nuclia/core';
+  import { FIELD_TYPE, Search } from '@nuclia/core';
   import { createEventDispatcher } from 'svelte';
+  import { getFieldUrl } from '../../core/stores/resource.store';
+  import { take } from 'rxjs';
 
   export let expanded = false;
   export let headerActionsWidth = 0;
@@ -18,6 +20,19 @@
 
   function clickOnTitle() {
     dispatch('clickOnTitle');
+  }
+
+  function openOrigin() {
+    if (!result.field) {
+      return;
+    }
+    getFieldUrl(result.field)
+      .pipe(take(1))
+      .subscribe((url) => {
+        if (url) {
+          window.open(url, 'blank', 'noreferrer');
+        }
+      });
   }
 </script>
 
@@ -34,7 +49,15 @@
       on:click={() => clickOnTitle()}>
       {result?.title}
     </h3>
+    {#if expanded && (result.field?.field_type === FIELD_TYPE.file || result.field?.field_type === FIELD_TYPE.link)}
+      <IconButton
+        icon={result.field.field_type === FIELD_TYPE.file ? 'download' : 'square-arrow'}
+        ariaLabel={$_('resource.source')}
+        aspect="basic"
+        on:click={openOrigin} />
+    {/if}
   </div>
+
   {#if expanded}
     <div class="header-actions">
       <slot />